% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday Function}
\usage{
birthday(x)
}
\arguments{
\item{x}{number of people in sample}
}
\value{
probability of two or more people with a common birthday
}
\description{
Birthday Function
}
\examples{
birthday(20:24)
}
