% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybinplot.R
\name{mybinplot}
\alias{mybinplot}
\title{Binary Bar plot Simulator}
\usage{
mybinplot(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{number of trials}

\item{p}{probability}
}
\value{
Bar plot
}
\description{
Binary Bar plot Simulator
}
\examples{
mybinplot(100,10,0.7)
}
