% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Poisson Distribution}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{lambda}{the Poisson rate parameter}
}
\value{
histogram of the sample mean, a barplot of the sample Y and  probablitiy function for the distribution.
}
\description{
Poisson Distribution
}
\examples{
mycltp(n=10,iter=10000)
}
