% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Curve function}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{limit}
}
\value{
a curve with the shaded area between the curve and x-axis along with the calculated area
}
\description{
Curve function
}
\examples{
myncurve(mu=10,sigma=5, a=6)
}
