% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Overbooking problem}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats in a flight}

\item{gamma}{probability a plane will truly be overbooked}

\item{p}{probablility of a "show"}
}
\value{
two Objective function Vs n plots, one for discrete and one for the continuous case, and a list of values containing nd, nc, N, p and gammma.
}
\description{
Overbooking problem
}
\examples{
ntickets(400,0.02,0.95)
}
