% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrapping Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{sample}

\item{fun}{function being bootstrapped}

\item{alpha}{Confidence interval}

\item{cx}{Text label size}

\item{...}{Graphing settings}
}
\value{
Confidence interval, function used, the sample and Xstat
}
\description{
Bootstrapping Function
}
\examples{
myboot2(10000,x=rnorm(25,mean=25,sd=10))
}
